[C64] Move SaveRam implementation to its own partial
